% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdplotr.R
\name{hdplotr}
\alias{hdplotr}
\title{Lazy hd plotting}
\usage{
hdplotr(
  x,
  y,
  xlab = "x",
  ylab = "y",
  main = NULL,
  width = 5,
  height = 3.5,
  units = "in",
  resolution = 250,
  mfrow = c(1, 1),
  name,
  format = "png",
  save = FALSE
)
}
\arguments{
\item{save}{}
}
\value{

}
\description{
Lazy hd plotting
}
