% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdplotr.R
\name{hdplotr}
\alias{hdplotr}
\title{Lazy hd plotting}
\usage{
hdplotr(
  x,
  y,
  xlab = "x",
  ylab = "y",
  main = NULL,
  width = 5,
  height = 3.5,
  units = "in",
  resolution = 250,
  mfrow = c(1, 1),
  name,
  format = "png",
  save = FALSE
)
}
\arguments{
\item{x}{x-axle}

\item{y}{y-axle}

\item{xlab}{label}

\item{ylab}{label}

\item{main}{label}

\item{width}{size}

\item{height}{size}

\item{units}{unit}

\item{resolution}{quality}

\item{mfrow}{more pictures}

\item{name}{save as}

\item{format}{format}

\item{save}{To save, or not to save}
}
\value{
A picture.
}
\description{
Lazy hd plotting
}
